데이터베이스 접속 구조화
각 테이블별 파이썬 파일을 제작하여 접속 방식을 좀더 보기 쉽게 만듬
