+++ Readme.md
@@ -7,6 +7,9 @@
 
 ## 서버 프로그램 작업 로그
 
+### 20170119
+>Database에 접속 하는 테이블에 접근하는 함수들을 각 테이블 별 파이썬 파일을 만들어서 구조화 시킴 , Useage라는 어색한 단어를 Availability라는 단어로 변경함
+
 ### 20170118-17
 >해당 매장의 모든 공지 조회 기능 추가
