+++ webService/BackEnd/Database/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ["CouponAvailabilityDatabase", "CouponDatabase", "CouponShapeDatabase", "CustomerLocationDatabase"
+    , "MileageLogDatabase", "NoticeDatabase", "OptimalStockDatabase", "ProductDatabase", "StoreInfoDatabase"
+    , "StoreRegisteredCustomerDatabase", "UploaderLocationDatabase", "CustomerInfoDatabase", "SalesVolumeDatabase"]
