+++ webService/BackEnd/StoreAndCustomerManager.py
@@ -1,7 +1,8 @@
 # -*- coding: utf-8 -*-
 # 사용자와 매장과의 관계에 관한 사항을 다루는 소스
 from django.http import HttpResponse
-import DatabaseManager
+import Database
+from Database import StoreRegisteredCustomerDatabase
 
 def AddNewCustomerToTargetStore(request) :
     """
@@ -11,7 +12,7 @@ def AddNewCustomerToTargetStore(request) :
     dbQuery = "INSERT INTO 매장등록 정보 VALUES( " + str(myUserId) + str(myStoreId) + "," + str(myUserId) + ", " + str(myStoreId) + " , false);"
     """
 
-    return HttpResponse(DatabaseManager.AddToStoreAsNewMember(request))
+    return HttpResponse(StoreRegisteredCustomerDatabase.AddToStoreAsNewMember(request))
 
 def BanCustomerFromTargetStore(request) :
     """
@@ -20,13 +21,13 @@ def BanCustomerFromTargetStore(request) :
     dbQuery = "UPDATE 매장등록 정보 SET 회원탈퇴여부 = true WHERE 고유등록번호 = "+ str(myUniqueId)  +";"
     """
 
-    return HttpResponse(DatabaseManager.DelMemberFromStore(request))
+    return HttpResponse(StoreRegisteredCustomerDatabase.DelMemberFromStore(request))
 
 def GetStoreAndCustomerRegisteredInfo(request) :
-    return HttpResponse(DatabaseManager.GetStoreAndCustomerRegisteredInfo(request))
+    return HttpResponse(StoreRegisteredCustomerDatabase.GetStoreAndCustomerRegisteredInfo(request))
 
 def SendNotificationMessage(request) :
-    return HttpResponse(DatabaseManager.SendNotificationMessage(request))
+    return HttpResponse(StoreRegisteredCustomerDatabase.SendNotificationMessage(request))
 
 def SendCoupon(request) :
-    return HttpResponse(DatabaseManager.SendCouponMessage(request))
+    return HttpResponse(StoreRegisteredCustomerDatabase.SendCouponMessage(request))
